Don't allow deletion of a destination if carpools are connected to it

+++ app/admin/views.py
@@ -230,7 +230,11 @@ def destinations_delete(uuid):
     delete_form = DeleteDestinationForm()
     if delete_form.validate_on_submit():
         if delete_form.submit.data:
-            # TODO Check to make sure no one is using the destination?
+            if dest.carpools:
+                flash("This destination cannot be deleted because "
+                      "some carpools still use it", 'error')
+                return redirect(url_for('admin.destinations_show', uuid=uuid))
+
             db.session.delete(dest)
             db.session.commit()
